Refactor endpoint SSRF audit helpers
